+++ src/test/java/com/sprint/mission/discodeit/repository/UserRepositoryTest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit;
+package com.sprint.mission.discodeit.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
